+++ tests/test_ctf.py
@@ -1,11 +1,12 @@
 from unittest.mock import patch
+
 from ebic.utils.auth import AuthUser
 
 
 @patch.object(AuthUser, "auth", return_value="admin", autospec=True)
 def test_get_admin(mock_user, client):
     """Get CTF data in a tomogram (request for admin)"""
-    resp = client.get("/ctf/1")
+    resp = client.get("/tomograms/1/ctf")
     assert resp.status_code == 200
     assert len(resp.json()["items"]) == 4
 
@@ -14,7 +15,7 @@ def test_get_admin(mock_user, client):
 def test_get_em_admin(mock_user, client):
     """Get CTF data in a tomogram (request for EM admin).
     Non-EM tomograms cannot exist."""
-    resp = client.get("/ctf/1")
+    resp = client.get("/tomograms/1/ctf")
     assert resp.status_code == 200
     assert len(resp.json()["items"]) == 4
 
@@ -22,7 +23,7 @@ def test_get_em_admin(mock_user, client):
 @patch.object(AuthUser, "auth", return_value="user", autospec=True)
 def test_get_user(mock_user, client):
     """Get CTF data in a tomogram belonging to user"""
-    resp = client.get("/ctf/2")
+    resp = client.get("/tomograms/2/ctf")
     assert resp.status_code == 200
     assert len(resp.json()["items"]) == 4
 
@@ -30,5 +31,5 @@ def test_get_user(mock_user, client):
 @patch.object(AuthUser, "auth", return_value="user", autospec=True)
 def test_get_forbidden(mock_user, client):
     """Try to get CTF data in a tomogram that does not belong to user"""
-    resp = client.get("/ctf/1")
+    resp = client.get("/tomograms/1/ctf")
     assert resp.status_code == 403